Extracted get-dmarc-record into separate function
