+++ lib/dmarc/verify.js
@@ -4,73 +4,7 @@ const dns = require('dns').promises;
 const punycode = require('punycode/');
 const psl = require('psl');
 const { formatAuthHeaderRow, getAlignment } = require('../tools');
-
-const resolveTxt = async (domain, resolver) => {
-    try {
-        let txt = await resolver(`_dmarc.${domain}`, 'TXT');
-        if (!txt || !txt.length) {
-            return false;
-        }
-
-        txt = txt.map(row => row.join('').trim()).filter(row => /^v=DMARC1\b/i.test(row));
-
-        if (txt.length !== 1) {
-            //no records or multiple records yield in no policy
-            return false;
-        }
-
-        return txt[0];
-    } catch (err) {
-        if (err.code === 'ENOTFOUND' || err.code === 'ENODATA') {
-            return false;
-        }
-        throw err;
-    }
-};
-
-const getDmarcRecord = async (domain, resolver) => {
-    let txt = await resolveTxt(domain, resolver);
-    let isOrgRecord = false;
-
-    if (!txt) {
-        let orgDomain = psl.get(domain);
-        if (orgDomain !== domain) {
-            // try org domain as well
-            txt = await resolveTxt(orgDomain, resolver);
-            isOrgRecord = true;
-        }
-    }
-
-    if (!txt) {
-        return false;
-    }
-
-    let parsed = Object.fromEntries(
-        txt
-            .split(';')
-            .map(e => e.trim())
-            .filter(e => e)
-            .map(e => {
-                let splitPos = e.indexOf('=');
-                if (splitPos < 0) {
-                    return [e.toLowerCase().trim(), false];
-                } else if (splitPos === 0) {
-                    return [false, e];
-                }
-                let key = e.substr(0, splitPos).toLowerCase().trim();
-                let val = e.substr(splitPos + 1);
-                if (['pct', 'ri'].includes(key)) {
-                    val = parseInt(val, 10) || 0;
-                }
-                return [key, val];
-            })
-    );
-
-    parsed.rr = txt;
-    parsed.isOrgRecord = isOrgRecord;
-
-    return parsed;
-};
+const getDmarcRecord = require('./get-dmarc-record');
 
 const verifyDmarc = async opts => {
     let { headerFrom, spfDomains, dkimDomains, resolver, arcResult } = opts;
